requirements.txt
pymoo==0.4.2.2
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.2
nltk==3.6.2
mysql-connector-python==8.0.26
